refactored models name
