+++ app/views/queries.py
@@ -2,7 +2,7 @@ from flask import Blueprint, render_template, request, redirect, url_for, jsonif
 from flask_login import current_user, login_required
 
 from .. import db
-from ..models.models import Projects, Experiments
+from ..models.model import Projects, Experiments
 from ..util.table_queries import ProjectsDataTable, ExperimentsDataTable
